--- server/src/main/java/com/soothee/dairy/dto/InputDairyDTO.java
@@ -1,7 +1,0 @@
-package com.soothee.dairy.dto;
-
-import java.util.List;
-
-public interface InputDairyDTO {
-    List<String> getCondIdList();
-}
