+++ server/src/main/java/com/soothee/member/controller/request/MemberDeleteRequest.java
@@ -1,4 +1,4 @@
-package com.soothee.member.dto;
+package com.soothee.member.controller.request;
 
 import com.soothee.custom.valid.ExistReferenceId;
 import io.swagger.v3.oas.annotations.media.Schema;
@@ -12,12 +12,12 @@ import java.util.List;
  * 회원 탈퇴
  * 1. 회원 일련번호 2. 선택한 탈퇴 사유 일련번호 리스트
  */
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Setter
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
-@Schema(description = "회원 탈퇴 DTO")
-public class MemberDelDTO {
+@Schema(description = "회원 탈퇴 요청 파라미터")
+public class MemberDeleteRequest {
     @NotEmpty(message = "회원 일련번호가 없습니다.")
     @Positive(message = "일련번호는 양수만 입력 가능합니다.")
     @Schema(description = "회원 일련번호")
