+++ server/src/main/java/com/soothee/member/controller/response/MemberResponse.java
@@ -1,0 +1,5 @@
+package com.soothee.member.controller.response;
+
+public interface MemberResponse {
+    Long getMemberId();
+}
