--- server/src/main/java/com/soothee/member/dto/MemberDTO.java
@@ -1,5 +1,0 @@
-package com.soothee.member.dto;
-
-public interface MemberDTO {
-    Long getMemberId();
-}
