+++ server/src/main/java/com/soothee/stats/controller/response/MonthlyContentDetail.java
@@ -1,9 +1,5 @@
-package com.soothee.stats.dto;
+package com.soothee.stats.controller.response;
 
-import com.soothee.common.constants.ContentType;
-import com.soothee.custom.exception.IncorrectValueException;
-import com.soothee.custom.exception.NullValueException;
-import com.soothee.custom.valid.SootheeValidation;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.PositiveOrZero;
@@ -16,11 +12,11 @@ import java.util.List;
  *  1. 한 달 동안 감사한/배운 일 작성 횟수
  *  2. 한 달 동안 작성한 모든 감사한/배운 일 정보
  */
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Setter
 @Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Schema(description = "한 달 동안 작성한 모든 감사한/배운 일 리스트 조회")
-public class MonthlyAllContentsDTO {
+public class MonthlyContentDetail implements StatsResponse {
     @NotEmpty(message = "한 달 동안 감사한/배운 일 작성 횟수가 없습니다.")
     @PositiveOrZero(message = "작성 횟수는 0을 포함한 양수만 입력 가능합니다.")
     @Schema(description = "한 달 동안 감사한/배운 일 작성 횟수")
@@ -30,8 +26,7 @@ public class MonthlyAllContentsDTO {
     @Schema(description = "한 달 동안 작성한 모든 감사한/배운 일 정보")
     private List<DateContents> contentList;
 
-    public MonthlyAllContentsDTO(Integer count, ContentType type) throws IncorrectValueException, NullValueException {
-        SootheeValidation.checkInteger(count, type);
+    public MonthlyContentDetail(Integer count) {
         this.count = count;
     }
 }
