+++ server/src/main/java/com/soothee/stats/controller/response/StatsResponse.java
@@ -1,0 +1,5 @@
+package com.soothee.stats.controller.response;
+
+public interface StatsResponse {
+    Integer getCount();
+}
