--- server/src/main/java/com/soothee/stats/dto/MonthlyContentsDTO.java
@@ -1,34 +1,0 @@
-package com.soothee.stats.dto;
-
-import com.soothee.common.constants.ContentType;
-import com.soothee.custom.exception.IncorrectValueException;
-import com.soothee.custom.exception.NullValueException;
-import com.soothee.custom.valid.SootheeValidation;
-import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.validation.constraints.NotEmpty;
-import jakarta.validation.constraints.PositiveOrZero;
-import lombok.*;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Setter
-@Getter
-@Schema(description = "한 달 동안 감사한/배운 일 요약")
-public class MonthlyContentsDTO {
-    @NotEmpty(message = "한 달 동안 작성 횟수가 없습니다.")
-    @PositiveOrZero(message = "작성 횟수는 0을 포함한 양수만 입력 가능합니다.")
-    @Schema(description = "한 달 동안 작성한 감사한/배운 일 횟수")
-    private Integer count;
-
-    @NotEmpty(message = "최고점 감사한/배운 일이 없습니다.")
-    @Schema(description = "한 달 중 가장 높은 점수를 준 날의 감사한/배운 일")
-    private DateContents highest;
-
-    @NotEmpty(message = "최저점 감사한/배운 일이 없습니다.")
-    @Schema(description = "한 달 중 가장 낮은 점수를 준 날의 감사한/배운 일")
-    private DateContents lowest;
-
-    public MonthlyContentsDTO(int count, ContentType type) throws IncorrectValueException, NullValueException {
-        SootheeValidation.checkInteger(count, type);
-        this.count = count;
-    }
-}
